:ambulance: Move reset history to HistoryController
